Fixing issues and attempting to run
