+++ questionbox/project/urls.py
@@ -16,10 +16,11 @@ Including another URLconf
 from django.contrib import admin
 from django.conf import settings
 from django.urls import include, path
+from . import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('question/', include('question.urls')),
+    path('questions/', include('questions.urls')),
 ]
 
 if settings.DEBUG:
